# Implement hierarchical queues for query scheduler