### What this PR does / why we need it:

This PR is the first step towards hierarchical queues within the query scheduler.
Hierarchical queues can be used to ensure query fairness across multiple actors within a single tenant, such as individual (human) users using the same tenant.

The change is described in the LID ["Query fairness across users within tenants"](https://grafana.com/docs/loki/next/lids/0003-queryfairnessinscheduler/).

* Extract metrics for `RequestQueue` into their own `metrics` field, so that they can extended and passed more easily in the future.
* Move mapping for tenant queues into its own data structure (ordered map) so its functionality can be reused. It also simplifies the code inside the `tenantQueues` struct.
* Add the `LeafQueue` struct that implements hierarchical queues. Sub-queues are dequeued in a round-robin manner using the same implementation of `Mapping` that is also used by the tenant queues.

### Special notes for your reviewer

Part of https://github.com/grafana/loki/pull/8585
Contains the changes from PR https://github.com/grafana/loki/pull/8722

**This PR does not yet make use of the new hierarchical queues, it only adds the relevant data structures. Wiring up will be done in a follow up PR.**

While the new `Mapping` data structure adds an unsignificant overhead to the time/op I consider the simplification an overall better tradeoff.

```console
$ benchstat scheduler-bench-old.txt scheduler-bench-new.txt 
name              old time/op    new time/op    delta
GetNextRequest-8    18.0µs ± 2%    20.5µs ± 3%  +13.89%  (p=0.008 n=5+5)
QueueRequest-8      56.6µs ± 0%    53.5µs ± 3%   -5.48%  (p=0.008 n=5+5)

name              old alloc/op   new alloc/op   delta
GetNextRequest-8    1.60kB ± 0%    1.60kB ± 0%     ~     (all equal)
QueueRequest-8      35.6kB ± 0%    29.8kB ± 0%  -16.37%  (p=0.008 n=5+5)

name              old allocs/op  new allocs/op  delta
GetNextRequest-8       100 ± 0%       100 ± 0%     ~     (all equal)
QueueRequest-8         819 ± 0%       806 ± 0%   -1.59%  (p=0.008 n=5+5)
```

#### Complexity of operations

| Operation | Function | Complexity |
| --------- | -------- | ---------- |
| Insert | `Put(key string, value *tenantQueue) bool` | `O(1)`|
| Select | `GetNext(idx QueueIndex) *tenantQueue` | `O(1)` (best case) `O(n)` (worst case) |
| Remove | `Remove(key string) bool` | `O(1)` |

**Checklist**
- [x] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [x] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/upgrading/_index.md`